+++ App.js
@@ -1,162 +1,10 @@
-import React, { useState } from "react";
-import { StatusBar } from 'expo-status-bar';
-import { 
-  StyleSheet, 
-  Text, 
-  Image, 
-  TextInput, 
-  View, 
-  Button,
-  FlatList,
-  SafeAreaView,
-  TouchableOpacity,
-} from 'react-native';
+import React from "react";
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-
-const DATA = [
-  {
-    id: "1",
-    title: "First Item",
-  },
-  {
-    id: "2",
-    title: "Second Item",
-  },
-  {
-    id: "3",
-    title: "Third Item",
-  },
-  {
-    id: "4",
-    title: "First Item",
-  },
-  {
-    id: "5",
-    title: "Second Item",
-  },
-  {
-    id: "6",
-    title: "Third Item",
-  },
-  {
-    id: "7",
-    title: "First Item",
-  },
-  {
-    id: "8",
-    title: "Second Item",
-  },
-  {
-    id: "9",
-    title: "Third Item",
-  },
-  {
-    id: "10",
-    title: "First Item",
-  },
-  {
-    id: "11",
-    title: "Second Item",
-  },
-  {
-    id: "12",
-    title: "Third Item",
-  },
-];
-
-const Item = ({ item, onPress, backgroundColor, textColor }) => (
-  <TouchableOpacity onPress={onPress} style={[styles.item, backgroundColor]}>
-    <Image
-    source={{
-      uri: 'https://reactnative.dev/docs/assets/p_cat2.png',}}
-      style={{ width: 20, height: 20 }}/>
-    <Text style={[styles.title, textColor]}>{item.title}</Text>
-  </TouchableOpacity>
-);
-
-
-function HomeScreen({navigation}) {
-  const [selectedId, setSelectedId] = useState(null);
-
-  const renderItem = ({ item }) => {
-    const backgroundColor = item.id === selectedId ? "#86AD09" : "gray";
-    const color = item.id === selectedId ? 'white' : 'black';
-
-    return (
-      <Item
-        item={item}
-        // onPress={() => setSelectedId(item.id)}
-        onPress={() => navigation.navigate('Details')}
-        backgroundColor={{ backgroundColor }}
-        textColor={{ color }}
-      />
-    );
-  };
-
-  return (
-    <SafeAreaView style={styles.container}>
-      <FlatList
-        data={DATA}
-        renderItem={renderItem}
-        keyExtractor={(item) => item.id}
-        extraData={selectedId}
-      />
-    </SafeAreaView>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    marginTop: StatusBar.currentHeight || 0,
-    backgroundColor: "white",
-  },
-  item: {
-    flexDirection: 'row',
-    padding: 20,
-    marginVertical: 8,
-    marginHorizontal: 20,
-  },
-  title: {
-    fontSize: 20,
-  },
-});
-
-
-
-// function HomeScreen({navigation}) {
-//   return (
-//     <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-//       <Text>Home Screen</Text>
-//       <Button
-//       title="Go to Details"
-//       onPress={() => navigation.navigate('Details')}/>
-//     </View>
-//   );
-// }
-
-
-function DetailsScreen({navigation}) {
-  return (
-    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-      <Text>Details Screen</Text>
-      <Button
-      title="Go to Details... again"
-      onPress={() => navigation.push('Details')}/>
-      <Button title="Go to Home" onPress={() => navigation.navigate('Home')} />
-      <Button title="Go back" onPress={() => navigation.goBack()} />
-      <Button
-        title="Go back to first screen in stack"
-        onPress={() => navigation.popToTop()}
-      />
-    </View>
-  );
-}
-
+import HomeView from "./ProyectData/Views/index"
+import DetailsScreen from "./ProyectData/Views/Details";
 
 const Stack = createNativeStackNavigator();
-
 
 function App() {
   return (
@@ -165,16 +13,20 @@ function App() {
         screenOptions={{
           headerStyle: {
             backgroundColor: '#1168AD'},
-          headerTintColor: 'black',
+          headerTintColor: 'white',
           headerTitleAlign: 'center',
-        }}
-      >
+        }}>
+
         <Stack.Screen
           name="Home"
-          component={HomeScreen}
-          options={{ title: 'Lista Productos' }}
-        />
-        <Stack.Screen name="Details" component={DetailsScreen} />
+          component={HomeView}
+          options={{ title: 'Lista Productos' }}/>
+
+        <Stack.Screen 
+        name="Details" 
+        component={DetailsScreen}
+        options={{ title: 'Details' }}/>
+
       </Stack.Navigator>
     </NavigationContainer>
   );
