+++ ProyectData/Views/Details.js
@@ -1,0 +1,35 @@
+import React, { useState } from "react";
+import { StatusBar } from 'expo-status-bar';
+import { 
+  StyleSheet, 
+  Text, 
+  Image, 
+  TextInput, 
+  View, 
+  Button,
+  FlatList,
+  SafeAreaView,
+  TouchableOpacity,
+} from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+
+function DetailsScreen({navigation}) {
+    return (
+      <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+        <Text>Details Screen</Text>
+        <Button
+        title="Go to Details... again"
+        onPress={() => navigation.push('Details')}/>
+        <Button title="Go to Home" onPress={() => navigation.navigate('Home')} />
+        <Button title="Go back" onPress={() => navigation.goBack()} />
+        <Button
+          title="Go back to first screen in stack"
+          onPress={() => navigation.popToTop()}
+        />
+      </View>
+    );
+}
+
+export default DetailsScreen;
+
